requirements.txt
absl-py==2.3.1
amqp==5.3.1
asgiref==3.9.1
astunparse==1.6.3
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.13.4
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
branca==0.8.1
bs4==0.0.2
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
decorator==5.2.1
deepface==0.0.93
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==65.10.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dlib @ file:///C:/Users/user/MyNewEnv/MyNewEnv/dlib-19.24.99-cp312-cp312-win_amd64.whl#sha256=20c62e606ca4c9961305f7be3d03990380d3e6c17f8d27798996e97a73271862
dnspython==2.8.0
drf-yasg==1.21.10
eventlet==0.40.3
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.18.0
fire==0.7.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
folium==0.20.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
geocoder==1.38.1
geographiclib==2.0
geojson==3.2.0
geolocator==0.1.1
geopandas==1.1.1
geopy==2.4.1
git-filter-repo==2.47.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.73.1
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
hyperlink==21.0.0
idna==3.10
imgkit==1.2.3
incremental==24.7.2
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
keras==3.10.0
kombu==5.5.4
libclang==18.1.1
lz4==4.4.4
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
msgpack==1.1.1
mtcnn==1.0.0
namex==0.1.0
numpy==1.26.4
oauthlib==3.3.1
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.16.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
prompt_toolkit==3.0.52
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pyogrio==0.11.1
pyOpenSSL==25.1.0
pyproj==3.7.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
ratelim==0.1.6
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==14.0.0
selenium==4.35.0
service-identity==24.2.0
setuptools==80.9.0
shapely==2.1.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
tf_keras==2.19.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
xyzservices==2025.4.0
zope.interface==7.2
